refactor(ui): separate memory settings controller
