+++ apps/desktop/src/main/__tests__/settings-contract-source-helpers.ts
@@ -18,6 +18,9 @@ const sourcePaths = [
   'appearance-settings-page.tsx',
   'web-search-settings-page.tsx',
   'memory-settings-page.tsx',
+  'use-memory-settings-controller.ts',
+  'memory-settings-sections.tsx',
+  'memory-entry-list.tsx',
   'memory-settings-labels.ts',
   'settings-error-copy.ts',
   'general-settings-page.tsx',
